Evaluate `:return-fn` compile-time
This fixes what I would consider to be a bug where stubadub would re-evaluate
the expression for `:return-fn` on every call to the stubbed function. This
behavior necessitates the use of a let or similar to create stateful stubs.
Stateful stuff isn't nice, but neither are stubs, so what are ya gonna do?